Inlined telegram decoders
