+++ src/main/scala/canoe/methods/chats/GetChatMember.scala
@@ -1,6 +1,5 @@
 package canoe.methods.chats
 
-import canoe.marshalling.CirceDecoders
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.{ChatId, ChatMember, InputFile}
@@ -23,7 +22,7 @@ object GetChatMember {
 
       def encoder: Encoder[GetChatMember] = deriveEncoder[GetChatMember].snakeCase
 
-      def decoder: Decoder[ChatMember] = CirceDecoders.chatMemberDecoder
+      def decoder: Decoder[ChatMember] = ChatMember.chatMemberDecoder
 
       def uploads(request: GetChatMember): List[(String, InputFile)] = Nil
     }
