+++ src/main/scala/canoe/methods/messages/SendAudio.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.ParseMode.ParseMode
@@ -55,7 +55,7 @@ object SendAudio {
 
       def encoder: Encoder[SendAudio] = deriveEncoder[SendAudio].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendAudio): List[(String, InputFile)] =
         List("audio" -> request.audio)
