+++ src/main/scala/canoe/methods/messages/SendContact.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.messages.TelegramMessage
@@ -44,7 +44,7 @@ object SendContact {
 
       def encoder: Encoder[SendContact] = deriveEncoder[SendContact].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendContact): List[(String, InputFile)] = Nil
     }
