+++ src/main/scala/canoe/methods/messages/SendDocument.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.ParseMode.ParseMode
@@ -45,7 +45,7 @@ object SendDocument {
 
       def encoder: Encoder[SendDocument] = deriveEncoder[SendDocument].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendDocument): List[(String, InputFile)] =
         List("document" -> request.document)
