+++ src/main/scala/canoe/methods/messages/SendGame.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.messages.TelegramMessage
@@ -37,7 +37,7 @@ object SendGame {
 
       def encoder: Encoder[SendGame] = deriveEncoder[SendGame].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendGame): List[(String, InputFile)] = Nil
     }
