+++ src/main/scala/canoe/methods/messages/SendLocation.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.messages.TelegramMessage
@@ -42,7 +42,7 @@ object SendLocation {
 
       def encoder: Encoder[SendLocation] = deriveEncoder[SendLocation].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendLocation): List[(String, InputFile)] = Nil
     }
