+++ src/main/scala/canoe/methods/messages/SendMessage.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.ParseMode.ParseMode
@@ -71,7 +71,7 @@ object SendMessage {
 
       def encoder: Encoder[SendMessage] = deriveEncoder[SendMessage].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendMessage): List[(String, InputFile)] = Nil
     }
