+++ src/main/scala/canoe/methods/messages/SendSticker.scala
@@ -1,11 +1,11 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.messages.TelegramMessage
 import canoe.models.{ChatId, InputFile, ReplyMarkup}
-import io.circe.generic.semiauto.deriveEncoder
+import io.circe.generic.semiauto._
 import io.circe.{Decoder, Encoder}
 
 /** Use this method to send .webp stickers. On success, the sent Message is returned.
@@ -38,7 +38,7 @@ object SendSticker {
 
       def encoder: Encoder[SendSticker] = deriveEncoder[SendSticker].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendSticker): List[(String, InputFile)] =
         List("sticker" -> request.sticker)
