+++ src/main/scala/canoe/methods/messages/SendVideo.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.ParseMode.ParseMode
@@ -53,7 +53,7 @@ object SendVideo {
 
       def encoder: Encoder[SendVideo] = deriveEncoder[SendVideo].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendVideo): List[(String, InputFile)] =
         List("video" -> request.video)
