+++ src/main/scala/canoe/methods/messages/SendVoice.scala
@@ -1,6 +1,6 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
+
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.ParseMode.ParseMode
@@ -49,7 +49,7 @@ object SendVoice {
 
       def encoder: Encoder[SendVoice] = deriveEncoder[SendVoice].snakeCase
 
-      def decoder: Decoder[TelegramMessage] = CirceDecoders.telegramMessageDecoder
+      def decoder: Decoder[TelegramMessage] = TelegramMessage.telegramMessageDecoder
 
       def uploads(request: SendVoice): List[(String, InputFile)] =
         List("voice" -> request.voice)
