+++ src/main/scala/canoe/methods/messages/StopPoll.scala
@@ -1,10 +1,9 @@
 package canoe.methods.messages
 
-import canoe.marshalling.CirceDecoders
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.{ChatId, InputFile, Poll, ReplyMarkup}
-import io.circe.generic.semiauto.deriveEncoder
+import io.circe.generic.semiauto._
 import io.circe.{Decoder, Encoder}
 
 /**
@@ -30,7 +29,7 @@ object StopPoll {
 
       def encoder: Encoder[StopPoll] = deriveEncoder[StopPoll].snakeCase
 
-      def decoder: Decoder[Poll] = CirceDecoders.pollDecoder
+      def decoder: Decoder[Poll] = deriveDecoder[Poll]
 
       def uploads(request: StopPoll): List[(String, InputFile)] = Nil
     }
