+++ src/main/scala/canoe/methods/users/GetUserProfilePhotos.scala
@@ -1,10 +1,9 @@
 package canoe.methods.users
 
-import canoe.marshalling.CirceDecoders
 import canoe.marshalling.codecs._
 import canoe.methods.Method
 import canoe.models.{InputFile, UserProfilePhotos}
-import io.circe.generic.semiauto.deriveEncoder
+import io.circe.generic.semiauto._
 import io.circe.{Decoder, Encoder}
 
 /** Use this method to get a list of profile pictures for a user. Returns a UserProfilePhotos object.
@@ -19,6 +18,7 @@ case class GetUserProfilePhotos(userId: Int,
                                )
 
 object GetUserProfilePhotos {
+  import io.circe.generic.auto._
 
   implicit val method: Method[GetUserProfilePhotos, UserProfilePhotos] =
     new Method[GetUserProfilePhotos, UserProfilePhotos] {
@@ -27,7 +27,7 @@ object GetUserProfilePhotos {
 
       def encoder: Encoder[GetUserProfilePhotos] = deriveEncoder[GetUserProfilePhotos].snakeCase
 
-      def decoder: Decoder[UserProfilePhotos] = CirceDecoders.userProfilePhotosDecoder
+      def decoder: Decoder[UserProfilePhotos] = deriveDecoder[UserProfilePhotos]
 
       def uploads(request: GetUserProfilePhotos): List[(String, InputFile)] = Nil
     }
