+++ src/main/scala/canoe/models/Animation.scala
@@ -9,7 +9,7 @@ package canoe.models
   * @param mimeType String Optional. MIME type of the file as defined by sender
   * @param fileSize Integer Optional. File size
   */
-case class Animation(fileId: String,
+final case class Animation(fileId: String,
                      thumb: Option[PhotoSize] = None,
                      fileName: Option[String] = None,
                      mimeType: Option[String] = None,
