+++ src/main/scala/canoe/models/Chat.scala
@@ -1,7 +1,23 @@
 package canoe.models
 
+import canoe.models.ChatType.ChatType
+import io.circe.Decoder
+import io.circe.generic.semiauto.deriveDecoder
+
 sealed trait Chat {
   def id: Long
+}
+
+object Chat {
+
+  implicit val chatDecoder: Decoder[Chat] = Decoder.instance[Chat] {
+    cursor => cursor.get[ChatType]("type").map {
+      case ChatType.Private => deriveDecoder[PrivateChat]
+      case ChatType.Group => deriveDecoder[GroupChat]
+      case ChatType.Supergroup => deriveDecoder[Supergroup]
+      case ChatType.Channel => deriveDecoder[Channel]
+    }.flatMap(_.tryDecode(cursor))
+  }
 }
 
 case class PrivateChat(id: Long,
